requirements.txt
fastapi
sqlmodel
psycopg2-binary
python-dotenv
supabase[embeddings]
uvicorn
celery
redis
python-docx
google-generativeai
